+++ pages/tags/[tagname].js
@@ -17,7 +17,7 @@ import { BsFillCalendar2DateFill } from 'react-icons/bs';
 
 const TagName = ({ post, tags, authors }) => {
   const blogsByTag = post.filter((p) => p.length > 0);
-  let options = { year: "numeric", month: "long", day: "numeric" };
+  let options = { year: "numeric", month: "short", day: "numeric" };
 
   return (
     <>
